Split classes into separate files.
