+++ python/lsst/ts/wep/task/DonutStamps.py
@@ -19,72 +19,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from dataclasses import dataclass
-
-import lsst.geom
-import lsst.afw.image as afwImage
-from lsst.meas.algorithms.stamps import AbstractStamp, StampsBase, readFitsWithOptions
-
-
-@dataclass
-class DonutStamp(AbstractStamp):
-    """Single donut stamp
-    Parameters
-    ----------
-    stamp_im : `lsst.afw.image.MaskedImageF`
-        The actual pixel values for the postage stamp
-    sky_position : `lsst.geom.SpherePoint`
-        Position of the center of the stamp.  Note the user
-        must keep track of the coordinate system
-    centroid_position : `lsst.geom.Point2I`
-        Position of the center of the stamp in pixels
-    detector_name : str
-        CCD where the donut is found
-    """
-
-    stamp_im: afwImage.maskedImage.MaskedImageF
-    sky_position: lsst.geom.SpherePoint
-    centroid_position: lsst.geom.Point2I
-    detector_name: str
-
-    @classmethod
-    def factory(cls, stamp_im, metadata, index):
-        """This method is needed to service the FITS reader.
-        We need a standard interface to construct objects like this.
-        Parameters needed to construct this object are passed in via
-        a metadata dictionary and then passed to the constructor of
-        this class.  If lists of values are passed with the following
-        keys, they will be passed to the constructor, otherwise dummy
-        values will be passed: RA_DEG, DEC_DEG.  They should
-        each point to lists of values.
-        Parameters
-        ----------
-        stamp : `lsst.afw.image.MaskedImage`
-            Pixel data to pass to the constructor
-        metadata : `dict`
-            Dictionary containing the information
-            needed by the constructor.
-        idx : `int`
-            Index into the lists in ``metadata``
-        Returns
-        -------
-        stamp : `DonutStamp`
-            An instance of this class
-        """
-        return cls(
-            stamp_im=stamp_im,
-            sky_position=lsst.geom.SpherePoint(
-                lsst.geom.Angle(metadata.getArray("RA_DEG")[index], lsst.geom.degrees),
-                lsst.geom.Angle(metadata.getArray("DEC_DEG")[index], lsst.geom.degrees),
-            ),
-            centroid_position=lsst.geom.Point2I(
-                metadata.getArray("CENT_X")[index], metadata.getArray("CENT_Y")[index]
-            ),
-            detector_name=metadata.getArray("DET_NAME")[index],
-        )
+from lsst.meas.algorithms.stamps import StampsBase, readFitsWithOptions
+from lsst.ts.wep.task.DonutStamp import DonutStamp
 
 
 class DonutStamps(StampsBase):
+    """
+    Holds a list of DonutStamp objects with helper functions.
+    """
 
     def _refresh_metadata(self):
         sky_positions = self.getSkyPositions()
@@ -100,23 +42,55 @@ class DonutStamps(StampsBase):
         self._metadata["DET_NAME"] = [det for det in det_names]
 
     def getSkyPositions(self):
+        """
+        Get the ra, dec coordinates in degrees of the DonutStamps
+
+        Returns
+        -------
+        list [lsst.geom.SpherePoint]
+            Ra, Dec locations of donuts
+        """
         return [s.sky_position for s in self._stamps]
 
     def getXY0Positions(self):
+        """
+        Get the corner position of the DonutStamp BBox locations.
+
+        Returns
+        -------
+        list [lsst.geom.Point2I]
+            Corner of BBox of DonutStamp lsst.afw.image.MaskedImage
+        """
         return [s.stamp_im.getXY0() for s in self._stamps]
 
     def getCentroidPositions(self):
+        """
+        Get the centroid positions of the DonutStamps in the original image.
+
+        Returns
+        -------
+        list [lsst.geom.Point2I]
+            X,Y pixel locations of center of DonutStamp images
+        """
         return [s.centroid_position for s in self._stamps]
 
     def getDetectorNames(self):
+        """
+        Get the detector name for each stamp.
+
+        Returns
+        -------
+        list [str]
+            Detector Names for each stamp
+        """
         return [s.detector_name for s in self._stamps]
 
     def append(self, item):
         """Add an additional stamp.
         Parameters
         ----------
-        item : `Stamp`
-            Stamp object to append.
+        item : DonutStamp
+            DonutStamp object to append.
         """
         if not isinstance(item, DonutStamp):
             raise ValueError("Objects added must be a DonutStamp object.")
@@ -127,12 +101,12 @@ class DonutStamps(StampsBase):
         """Extend DonutStamps instance by appending elements from another instance.
         Parameters
         ----------
-        stamps_list : `list` [`DonutStamp`]
+        stamps_list : list [DonutStamp]
             List of DonutStamp object to append.
         """
         for s in stamp_list:
             if not isinstance(s, DonutStamp):
-                raise ValueError("Can only extend with DonutStamp objects")
+                raise ValueError("Can only extend with DonutStamp objects.")
         self._stamps += stamp_list
 
     @classmethod
@@ -140,11 +114,11 @@ class DonutStamps(StampsBase):
         """Build an instance of this class from a file.
         Parameters
         ----------
-        filename : `str`
+        filename : str
             Name of the file to read
         Returns
         -------
-        object : `DonutStamps`
+        DonutStamps
             An instance of this class
         """
         return cls.readFitsWithOptions(filename, None)
@@ -154,13 +128,13 @@ class DonutStamps(StampsBase):
         """Build an instance of this class with options.
         Parameters
         ----------
-        filename : `str`
+        filename : str
             Name of the file to read
-        options : `PropertyList` or `dict`
+        options : PropertyList or dict
             Collection of metadata parameters
         Returns
         -------
-        object : `DonutStamps`
+        DonutStamps
             An instance of this class
         """
         stamps, metadata = readFitsWithOptions(filename, DonutStamp.factory, options)
